feat: include shr.id, shr.time, shr.price, r.id, and r.name in showtime output

+++ entity/src/showtime_room.rs
@@ -8,7 +8,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub time: DateTime,
-    pub price: Decimal,
+    pub price: i32,
     pub room_id: Uuid,
     pub showtime_id: Uuid,
 }
